fixed wrong interpretation of image type

+++ createThumbnail.py
@@ -13,9 +13,11 @@ arg2 = sys.argv[2]
 arg3 = sys.argv[3]
 
 
-def generateThumbnail(path=arg2):
+def generateThumbnail(path=arg2, empty=arg3):
     imagename = ntpath.basename(path).rsplit(".", 2)
-    if imghdr.what(path) == 'png' or 'jpeg' or 'tiff':
+    img = Image.open(path)
+    print(imagename[1])
+    if imagename[1] == "jpg" or imagename[1] == "tif" or imagename[1] == "tiff" or imagename[1] == "png":
         thumb = Image.open(path)
         thumb = thumb.convert('RGB')
     else:
@@ -27,14 +29,15 @@ def generateThumbnail(path=arg2):
     print(savepath)
     imageio.imsave(savepath, thumb)
 
-def delete(path=arg2):
+def delete(path=arg2, empty=arg3):
     imagename = ntpath.basename(path).rsplit(".", 2)
     deletepath = ntpath.dirname(path) + "/thumbnail/" + imagename[0] + "_thumb.jpg"
     deleteeditedpath = ntpath.dirname(path) + "/edited/" + imagename[0] + ".tiff"
     print(deletepath)
     os.remove(path)
     os.remove(deletepath)
-    os.remove(deleteeditedpath)
+    if os.path.exists(deleteeditedpath):
+        os.remove(deleteeditedpath)
 
 def prepareDownload(path=arg2, type=arg3):
     imagename = ntpath.basename(path).rsplit(".", 2)
